ref-filter: rename `objectname` related functions and variable
In previous commits, we prepared some `objectname` related functions
for more generic usage. So, same logic can be used for `tree` and
`parent` atom.

But the name of some fuctions and variables may mislead someone.
For ex: Function `objectname_atom_parser()` implies that it is
for atom `objectname`.

Let's rename all such functions and variables.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Heba Waly <heba.waly@gmail.com>
Signed-off-by: Hariom Verma <hariom18599@gmail.com>